% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmeanshfit.R
\name{meanshift}
\alias{meanshift}
\title{Mean Shift Segmentation image data}
\usage{
meanshift(
  img,
  radiusS = 5,
  radiusR = 4.5,
  minDensity = 200,
  speedUp = 2,
  scale = TRUE
)
}
\arguments{
\item{img}{A 3D array [height, width, channels] representing a color image}

\item{radiusS}{Numeric scalar. Spatial bandwidth. Default 5.}

\item{radiusR}{Numeric scalar. Range/color bandwidth. Default 4.5.}

\item{minDensity}{Integer. Minimum density for clusters. Default 200.}

\item{speedUp}{Integer. Optional speed-up level. Default 2.}

\item{scale}{Boolean. Scales image values to 0-255. Default TRUE}
}
\value{
Returns an array of the same shape as the input
}
\description{
Performs mean shift segmentation on multi-channel images.
Input can be a 3D array [height, width, channels]
}
