% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshift_main.R
\name{segment}
\alias{segment}
\title{Segment an image using Mean Shift algorithm}
\usage{
segment(
  image,
  spatial_radius,
  range_radius,
  min_density,
  speedup_level = SPEEDUP_HIGH
)
}
\arguments{
\item{image}{Input image (2-D or 3-D array).}

\item{spatial_radius}{Spatial radius of the search window (integer).}

\item{range_radius}{Range radius of the search window (double).}

\item{min_density}{The minimum point density of a region in the segmented image (integer).}

\item{speedup_level}{Filtering optimization level for fast execution (default: high).}
}
\value{
A list containing:
  \item{segmented}{Image where the color of the regions is the mean value of the pixels belonging to a region.}
  \item{labels}{2-D array where a pixel value corresponds to the region number the pixel belongs to.}
  \item{nb_regions}{The number of regions found by the mean shift algorithm.}
}
\description{
Segment an image using Mean Shift algorithm
}
