% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshift.R
\docType{class}
\name{Segmenter-class}
\alias{Segmenter-class}
\alias{Segmenter}
\title{Segmenter class using the mean shift algorithm to segment images}
\description{
Segmenter class using the mean shift algorithm to segment images
}
\section{Methods}{

\describe{
\item{\code{segment(image)}}{Segment the input image (color or grayscale).}
}}

