% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmeanshfit.R
\name{meanshift}
\alias{meanshift}
\title{Mean Shift Segmentation on Raster or Array Images}
\usage{
meanshift(img, radiusS = 5, radiusR = 4.5, minDensity = 200, speedUp = 2)
}
\arguments{
\item{img}{A `terra::rast` object or 3D array [height, width, channels].}

\item{radiusS}{Numeric scalar. Spatial bandwidth. Default 5.}

\item{radiusR}{Numeric scalar. Range/color bandwidth. Default 4.5.}

\item{minDensity}{Integer. Minimum density for clusters. Default 200.}

\item{speedUp}{Integer. Optional speed-up level. Default 2.}
}
\value{
If input is `terra::rast`, returns segmented raster; if array, returns 3D array.
}
\description{
Performs mean shift segmentation on multi-channel images.
Input can be a `terra::rast` or a 3D array [height, width, channels].
}
